(api) Add customer existence check on deletion

+++ api/src/domain/customers/use-cases/delete-customer.ts
@@ -1,5 +1,6 @@
 import { inject, injectable } from 'tsyringe'
 import { CustomersRepository } from '../repositories/customers'
+import { CustomerNotFound } from '../errors/customer-not-found'
 
 interface DeleteCustomerUseCaseRequest {
   id: string
@@ -13,6 +14,10 @@ export class DeleteCustomerUseCase {
   ) {}
 
   async execute({ id }: DeleteCustomerUseCaseRequest): Promise<void> {
+    const customer = await this.customersRepository.findById(id)
+
+    if (!customer) throw new CustomerNotFound(id)
+
     return this.customersRepository.delete(id)
   }
 }
